Rename `traces` feature to the more standard `trace`
