+++ opentelemetry-proto/src/transform/mod.rs
@@ -3,8 +3,8 @@ pub mod common;
 #[cfg(feature = "metrics")]
 pub mod metrics;
 
-#[cfg(feature = "traces")]
-pub mod traces;
+#[cfg(feature = "trace")]
+pub mod trace;
 
 #[cfg(feature = "logs")]
 pub mod logs;
