+++ opentelemetry-proto/src/transform/trace.rs
